refactor: rename "noop" package to "collection"
